moved database file out of config and renamed it
